requirements.txt
# Core LangChain
langchain>=0.1.0,<0.3.0
langchain-community>=0.0.10,<0.3.0
langchain-openai>=0.0.5,<0.3.0
langchain-experimental>=0.0.50,<0.3.0

# Vector stores and embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Document processing
pypdf>=3.17.0
unstructured>=0.10.30,<0.12.0
python-multipart>=0.0.6

# Web APIs and tools
requests>=2.31.0
google-search-results>=2.4.2
tavily-python>=0.3.0

# Web framework and UI
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
streamlit>=1.29.0
gradio>=4.13.0

# Database and storage
sqlalchemy>=2.0.25
alembic>=1.13.1

# Evaluation and monitoring
datasets>=2.16.0
evaluate>=0.4.1
wandb>=0.16.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0,<3.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.1.0
tqdm>=4.66.0
rich>=13.7.0

# Development
pytest>=7.4.0
black>=23.12.0
flake8>=7.0.0